1、device 配置初始化 2、Modbus 通过写超时来识别未被配置的Channel。超时后断开远端 3、输入端口支持反向 4、修改UI中，Device界面和Channel界面的问题

+++ src/main/java/com/rytec/rec/node/HTKL/HT_C1.java
@@ -77,7 +77,7 @@ public class HT_C1 extends NodeAnalog implements NodeInterface {
         ByteBuf payload = (ByteBuf) msg.payload;
 
         int val = payload.getShort(3);
-        rst.value = val * nodeRuntimeBean.nodeConfig.pA + nodeRuntimeBean.nodeConfig.pB;
+        rst.value = (Float) (val * nodeRuntimeBean.nodeConfig.pA + nodeRuntimeBean.nodeConfig.pB);
         payload.release();
     }
